67. slot
