+++ Juyoung/vue-project/db.json
@@ -120,18 +120,6 @@
       "completed": false,
       "body": "",
       "id": 28
-    },
-    {
-      "subject": "11",
-      "completed": false,
-      "body": "",
-      "id": 29
-    },
-    {
-      "subject": "111",
-      "completed": false,
-      "body": "",
-      "id": 30
     }
   ]
 }
