+++ Juyoung/vue-project/src/components/Modal.vue
@@ -5,30 +5,17 @@
         <div class="modal-content">
           <div class="modal-header">
             <h5 class="modal-title">
-              Delete Todo
+              <slot name="title"></slot>
             </h5>
             <button type="button" class="close">
               <span @click="onClose">&times;</span>
             </button>
           </div>
           <div class="modal-body">
-            Are you sure tou want to delete the todo?
+            <slot name="body"></slot>
           </div>
           <div class="modal-footer">
-            <button 
-              type="button" 
-              class="btn btn-secondary" 
-              @click="onClose"
-            >
-              Close
-            </button>
-            <button 
-              type="button" 
-              class="btn btn-danger"
-              @click="onDelete"
-            >
-              Delete
-            </button>
+            <slot name="footer"></slot>
           </div>
         </div>
       </div>
@@ -42,13 +29,9 @@ export default {
     const onClose = () => {
       emit('close');
     }
-    const onDelete = () => {
-      emit('delete');
-    }
 
     return {
       onClose,
-      onDelete
     }
   }
 }
