+++ Juyoung/vue-project/src/components/TodoList.vue
@@ -40,7 +40,7 @@
 
 <script>
 import { useRouter } from 'vue-router';
-import Modal from '@/components/Modal.vue';
+import Modal from '@/components/DeleteModal.vue';
 import { ref } from 'vue';
 
 export default {
